transcribio-backend/flask-backend/requirements.txt
Flask
gunicorn
ffmpeg-python
requests
multi-rake
google-cloud-speech
firebase-admin
